Connect to orders state (#43)
* Override hook swap callback

* Override and delegate

* Provide mock implementation

* Custom redux store (#31)

* copy state/index to custom

* create sample of Operator actions

* create sample of Operator reducer

* add operator reducer to state

* create sample of Operator hooks

* changeto absolute import { AppDispatch, AppState } from 'state/'

* change to absolute import store from 'state/'

* minor organisational changes

* from 'state/' -> from 'state'

* special var for UNISWAP_REDUCERS

* simplify types

* fix import error


* Print useful trading information

* Disable token warning

* Disable token warning

* Add actions

* Add hook for adding pending orders

* Add redux state for orders

* Add expiration and update executed order

* Simplify order params

* Add logic for posting the order

* Rename param

* Fix some old type name

* Add deadline to logged params

* Do not check immutable changes, is too slow (#38)

* Disable token warning (#39)

* Make types closer to the future orders PR

* Mock signing transaction, call API, and update state

* Remove unnecesary things in the reducer

* Change valid to to number

* Fix how the buyToken is extracted

* Connect to orders state

* Re-delete code that a bad merge added

* Remove unsed imports

* Simplify expression

* Reapply commit after bad merge
