+++ src/custom/state/operator/hooks.ts
@@ -1,21 +1,8 @@
-import { ChainId } from '@uniswap/sdk'
-import { useCallback } from 'react'
 import { useDispatch, useSelector } from 'react-redux'
 
 import { AppDispatch, AppState } from 'state'
-import { updateTip, clearTip, addPendingOrder, OrderID, Order, Tip } from './actions'
+import { updateTip, clearTip, Tip } from './actions'
 import { OperatorState } from './reducer'
-
-interface AddPendingOrderParams extends GetRemoveOrderParams {
-  order: Order
-}
-
-interface GetRemoveOrderParams {
-  id: OrderID
-  chainId: ChainId
-}
-
-type AddOrderCallback = (addOrderParams: AddPendingOrderParams) => void
 
 interface AddTipParams extends ClearTipParams {
   tip: Tip
@@ -26,11 +13,6 @@ interface ClearTipParams {
 
 type AddTipCallback = (addTokenParams: AddTipParams) => void
 type ClearTipCallback = (clearTokenParams: ClearTipParams) => void
-
-export const useAddPendingOrder = (): AddOrderCallback => {
-  const dispatch = useDispatch<AppDispatch>()
-  return useCallback((addOrderParams: AddPendingOrderParams) => dispatch(addPendingOrder(addOrderParams)), [dispatch])
-}
 
 export const useTip = (tokenAddress: string): Tip | undefined => {
   const { tipsMap } = useSelector<AppState, OperatorState>(state => state.operator)
